Update ButtonTileApp.java
so mr d said to js do tile pane then work on overlapping when get back...I voided that. stack pane based on what I read seems to have the ability to overlap them. I used for loops for the buttons cus some of the tutorials did, others just made each one. I think for loops may make it easier if u plan on adding a bunch of buttons. idk if this looks how I want it to or even displays cus, yea, no java fx, but I dont think any syntax errors and I think the logic is fine except to the black keys. either it it'll place where I want it to, which is the first black second black no black third black 4th black 5th black, and the empty region all be the rest, or it'll just place in order. prettier made the code look good and I copied that bit where each button prints when it clicks. I dont know if I needed Java.util.scanner tho. oopsies. we still need to make it so the buttons display their keys, move or change color when pressed, and obviously map to a sound. also may need to add hgap