white keys work maybe? ButtonTileApp.java
When the black keys dont need to fill a space, it puts a region. I thought they were automatically transparent, I guess not. Idk if this will fix it, they could also need to have insets so that the black and white buttons dont overlap, but we can try that if this doesn't work. Otherwise well need to use a different box/pane.